Fix pylint R1714 (Issue #124)
IV_Swinger.py:1016:19: R1714: Consider merging these comparisons with "in" to 'mwp_num not in (point_num, point_num + 1)' (consider-using-in)
IV_Swinger2_gui.py:1970:15: R1714: Consider merging these comparisons with "in" to 'rc in (RC_ZERO_ISC, RC_ZERO_VOC, RC_ISC_TIMEOUT)' (consider-using-in)

consider-using-in (R1714) Consider merging these comparisons with "in" to %r To check if a variable is equal to one of many values,combine the values into a tuple and check if the variable is contained "in" it instead of checking for equality against each of the values.This is faster and less verbose.

Made recommended changes in both python and python3 files.