Remove IV_Swinger_test.py (Issue #124)
The IV_Swinger_test.py module has several pylint issues. Instead of fixing them, this module is removed (both from python and python3) since it is specific to IV Swinger 1 and was really just an example of how to create a class that is derived from the IV_Swinger class.