Disable pylint wrong-import-position (Issue #124)
The workaround for the Matplotlib crash in commit https://github.com/csatt/IV_Swinger/commit/2813c76bec1898218f9dbbed53e53a4262c153f5 works only if the workaround code is inserted between the import of matplotlib and the import of numpy. This causes pylint to generate a wrong-import-position warning for all imports after the workaround code. The only way around that is to disable this pylint check. This is in the Python 2 code only.