Log runtime OS version (Issue #125)
This started as a simple change, but I realized that most of the logging in the log_python_and_tcl_tk_versions() method in the GUI class really should be in the log_initial_debug_info() method in the IVS2 class, and the determination of most of the package versions actually belongs in IV_Swinger.py since that is where the imports are.

This commit does the following:
  - IV_Swinger2_gui.py:
     Change log_python_and_tcl_tk_versions() to simply
     log_tcl_tk_version(), removing the logging of the Python version
     and all of the Python package versions.
  - IV_Swinger.py:
     Set IV_Swinger class attributes with the values of:
       - OS version
       - Python version
       - Matplotlib version
  - IV_Swinger_PV_model.py:
     Get SciPy version into global namespace with import
  - IV_Swinger2.py
     Set IV_Swinger2 class attributes with the values of:
       - PySerial version
       - Pillow version
     Add the following logging to log_initial_debug_info():
       - OS version
       - Python version
       - Matplotlib version
       - SciPy version
       - PySerial version
       - Pillow version

Additionally, the conditional imports in IV_Swinger.py were changed to eliminate the copied lines. And in the Python 2 version, the workaround for the matplotlib crash on Python 2.7.18 is moved here from IV_Swinger2_gui.py, which no longer imports matplotlib.