Fix Windows breakage
Fix the problem caused by the commit for Issue #125.

Instead of using os.uname() to determine whether to import the Adafruit modules, just use try/except. This is better because RPi isn't the only ARM-based machine. Future Macs are even going to be ARM-based.

Instead of os.uname() to determine if processor is arm6, use platform.machine().