Fix double title in log file dialog (Mac)
No Issue filed. The title in the log file dialog is repeated when the Python 2 code is run with 2.7.18 or when the Python 3 code is run (Mac only). It is not duplicated when the Python 2 code is run with the Mac system Python. It is likely that it is really the Tcl/Tk version that is the determining factor, however.

In any case, the solution is to not set both the "title" and "message" option for tkfiledialog.askopenfilename(), but rather set only the "message" option if the code is running on Mac.

For some reason, this doesn't seem to be necessary for tkfiledialog.askdirectory(), which is used in get_copy_dest() and change_folder(). For consistency, and in case those change too in the future, the same change is made there too.