Fix loop mode checkbutton not updating
No Issue filed. Noticed in regression testing that when the Rate Limit and Save Results checkbuttons are set and then clicked to clear them, the checkmark doesn't clear until the mouse is moved. This is running on Mac with Python 2.7.18 or Python 3.8.5. It doesn't happen with Mac system Python. The difference may actually be the Tcl/Tk version rather than Python. It doesn't happen on Windows.

Solution is to add update_idletasks() in update_loop_rate_limit() and update_loop_save_results(). Also added to update_loop_mode() even though that checkbutton doesn't have the problem.