Fix ungraceful app exit (Issue #132)
Two changes are made:
    - IV_Swinger2.py, IV_Swinger2_plotter class: Add assert for existence of CSV file(s) before invoking CsvFileProcessor.
    - Change the sys.exit(-1)'s to return [] in IV_Swinger_plotter.py

This means non-existence (or incorrect formatt) of a data points CSV file causes an exception. This isn't super user-friendly because the user has to look at the log file to know what was wrong. But it is much better than sys.exit(-1) and isn't worth the coding to generate a dialog, given the unlikelihood of occurrence.