Fix Overlay pane in Results Wizard (Issue #133)
The geometry management in the add_overlay_widgets() method was poor. The three buttons were packed into a ttk.Frame (overlay_widget_buttonbox). Between the Cancel and Finish buttons there was a dummy ttk.Label (overlay_button_pad) to make a gap. The buttons had specified widths (determined by trial-and-error) to make the buttonbox the same width as the enclosing ttk.Frame (overlay_widget_box). It appears that the hardcoded button widths (and or pad width) are not correct for the current Pythons (or more likely, the current Tcl/Tk).

This is now done correctly. The overlay_button_widgetbox is placed into overlay_widget_box using sticky=(E, W) instead of sticky=(W). This "stretches" it to the same width as the parent overlay_widget_box (assuming the buttons fit). Now the Finish button is placed with pack(side=RIGHT) and the gap between the Cancel and Finish buttons is automatic. The buttons have their natural sizes (i.e. width= removed) and the pad is removed..