Remove hardcoded geometry values for simulator help dialog
No Issue filed. The simulator help dialog had a fixed min_height of 360 and max_height of 2000. This is changed so the min_height is half of the parent window's height and the max is the full screen height.

For consistency, the default max_height in the GUI set_dialog_geometry() method is changed from winfo_height to winfo_screenheight.

Also fixed a whitespace flake8 issue in IV_Swinger2_gui.py.