Use winfo_x() and winfo_width() instead of parsing geometry() (Issue #159)
Since Issue #159 hasn't been reproduced, it's hard to say what the cause was. But the failure was parsing the X value from the string returned by geometry(), which should be in the format: (width)x(height)+(X)+(Y)

Parsing the X value and width from geometry() is not necessary. These values are available from the winfo_x() and winfo_width() functions respectively. This may just change the synptom of the problem, but it should be more robust.