Catch exception when canceling overlay (Issue #162)
Add method rm_overlay_dir(), which catches the OSError exception on the call to shutil.rmtree and gives the user an opportunity to close the PDF viewer.  Although the exception is either WindowsError or PermissionError, OSError catches it.

One other small change in this commit is to display an error dialog when the send2trash() call fails with OSError in the delete_selected() method. Previously, this generated a message in the log file, but did not generate an error dialog.