Revamp SSR current calibration (Issue #163)
The algorithm for measuring the SSR current calibration current is changed to have two passes of measurements to average. The stability is evaluated based on the difference between these two averages. Previously, the stability was determined by how far out of whack the highest and lowest values seen were. This was too sensitive, and resulted in many false positives.

The accuracy should be improved slightly too. The second pass discards values that are more than +-0.5% from the average of the first pass. Also, the average of the second pass uses floating point and is rounded to the nearest integer (long). Previously, the average was an integer, so it could be off by nearly one ADC increment. Now the max rounding error is 0.5 ADC increment.