Set hdd_output_dir to None in request_adv_calibration_vals (Issue #164)
The request_adv_calibration_vals() method requests an IV curve from the Arduino, but it doesn't save the results anywhere. The receive_data_from_arduino() method writes the sensor values to the run_info file. But in this case, there shouldn't be a run_info file to write it to. The problem reported in Issue #164 occurs because the hdd_output_dir property is still pointing to the directory of the last run, which does have a run_info file.

The problem is fixed simply by setting the hdd_output_dir property to None (which its initial value) before calling receive_data_from_arduino(). However, the property has a sanity check that also needed to be modified in order to allow it to be set to None.