Add option to use equation #5 (Issue #167)
With equation #5 fixed, it raises the question: can it now be used?

This commit adds the option to use equation #5. The "use_eq5" property of the PV_model class defaults to False, but if it is set to True, then equation #5 is used.

The answer to the question is that it still causes some models to fail that work with the 4 equations, and it slows down modeling in all cases with no noticeable benefits. But it is nice to retain the ability to use all five equations for future experimentation.