Fix huge PDF problem (Issue #152 reopen)
Issue #152 was reopened because it was noticed that with the previous fix, the PDF files are huge and take a long time to generate.

This commit removes the previous fix (matplotlib.rcParams['pdf.fonttype'] = 42) and instead adds a wrapper around the plot_with_pyplot() method. This wrapper catches the exception "RuntimeError: TrueType font is missing table" and retries the call to plot_with_pyplot() with the font changed to the default. It also filters the annoying warnings that are generated in this case.

Also modified the help text for the Plotting Preferences tab of the GUI to remove mention of some fonts (e.g. Wingdings) requiring the application to be restarted. Wingdings work now and I can't find others that require the app to be restarted. Strangely, this seems to have been something wrong with my computer that "got better". The v2.6.0 released app even works now.