Enhance PV model to use equation #5 (Issue #168)
IV_Swinger_PV_model.py is modified with the following:
   - DEFAULT_RSH_GUESSES (formerly DEFAULT_RSH_VALS) now has a second value: 100 (ohms)
   - The test_all_five_parms function is now just test_parms, and the former test_parms function is now test_first_four_parms
   - An additional loop is added to the find_parms function to first try using equation 5 and only run without it if that fails
   - Added a PV cell spec to create_pv_spec_file()
   - Added measurement of time to execute run() method
   - Removed use_eq5 property
   - Updated all comments and docstrings to reflect the modifications

The reason that the find_parms function has the loop to not use equation #5 is that it was found that there are still some cases that coverge only when a hardcoded Rsh value is used (e.g. Canadian Solar CS3W-415P and Q.PEAK DUO-G5 330.)

The IV_Swinger2.py module has one small modification to include the model execution time in the plot legend, after the parameter values.