Move plot_and_view_modeled_curve() into main() (Issue #172)
The plot_and_view_modeled_curve() function is moved into main() as a local function.  This makes the "boilerplate" main() call much closer to being boilerplate. It still has the import of IV_Swinger2. This isn't really necessary; it could be at the top of the file, as normal. However, pylint flags this as a cyclic import. The cyclic import doesn't appear to cause a problem, but just in case, it seems safer to not import IV_Swinger2 unless the module is being run standalone. The pylint warning is disabled.