Update design doc to 1.2notoc
This commit is for the Word version of the design doc only (no PDF). This version of the doc has no TOC, TOF or TOE. What it does have is change bars. The file is corrupted in a way that prevents a PDF from being generated from it. Even though the TOC/TOF/TOE are removed, some vestige of them remains that is "poisoning" the document. The only way (I have found) to fix it is to start with a new blank document and copy/paste everything from this one (up to the final paragraph mark) into the new one. Among other things, this loses the change bars. The only reason for comitting this version is to have a record of the change bars.