Update design doc to version 1.2
This commit is for version 1.2 of the Word and PDF versions of the design doc. There are no change bars due to a file corruption problem. See the previously checked in Word doc (which is corrupted) for change bars.

The changes reflect changes made to the design since the 1.1 version of the doc. The primary ones are based on these issues:

   - #100 RFE: Use Arduino internal 1.1V bandgap to measure Vref enhancement
   - #104 Enhancement: simulator for scaled component selection enhancement
   - #105 Enhancement: add ability to change plotting font enhancement
   - #111 Enhancement: add support for PV reference curves enhancement
   - #121 Enhancement: Port to Python 3