Fix minor debug feature issue
No Issue was filed for this.

There is a debug feature that skips Isc polling if max_isc_poll is negative. For some reason, adc_i_val_prev_prev was set to 2048 in the code that handles this. Since that value is used for isc_adc, the Isc point has an ADC value of 2048 regardless of the measured points. adc_i_val_prev_prev is now set to adc_i_val instead.