Optimize Isc measurement for SSR/FET hardware (Issue #181)
Changed the algorithm that determines when Isc is stable so that when the hardware uses SSRs or FETs, the polling ends when the current and voltage stop changing (when SSR3/FET3 is on). Previously, when this condition occurred, the polling would continue, using the EMR algorithm. But this always resulted in an Isc point further away from the axis, meaning a larger extrapolation. The tricky part is that the code doesn't know if the hardware is EMR or SSR/FET. But the EMR conditions won't occur on SSR/FET hardware and the SSR/FET conditions won't happen on EMR hardware, so the code just looks for both of them in parallel, and the polling naturally terminates at the correct time for the given hardware.

Updated version number to 1.4.4