Add READ_ADC command for debug with serial monitor (Issue #184)
Added a new "config" (misnomer in this case) command to simply read both channels of the ADC and print their values. It has a count parameter that specifies the number of times to repeat the reads in a loop, once every 0.5 seconds. This is for debugging. It is intended to be invoked from the Serial Monitor of the Arduino application. For example, typing "Config: READ_ADC 20" in the command window of the serial monitor, and hitting the Send button will read both ADC channels 20 times over 10 seconds, printing the values in the monitor window.