Add remote command feature (Issue #196)
This commit adds the Remote Command feature. It uses the ZeroMQ (aka ZMQ, 0MQ) messaging library. This allows client programs to be written in various languages. The basic REQ/REP (client/server) messaging pattern is used.

The server is implemented as part of the GUI. This was a design decision (the other choice being to implement it at the lower IV_Swinger2.py level that controls the IV Swinger 2 hardware.) This means the remote commands are handled by the GUI, and the effects of the commands are visible on the GUI. Namely, when an IV curve is initiated remotely, the curve is displayed on the local screen just as if the user had clicked the "Swing!"button.

A new tab has been added to the Preferences dialog to configure the Remote Command feature. This tab is used to enable/disable remote command monitoring, to specify a TCP port to communicate over, and to specify the polling interval. If the "instances" feature is used, each instance is configured independently and can receive and execute remote commands concurrently with the other instances. Like other preferences, the choices on the Remote Command tab are saved in the Config file so they persist across runs.

The only remote command implemented in this commit is the "Swing" command. This command triggers the measurement of an IV curve and returns a summary of the results to the requester. If the command succeeds, the Isc, Voc, and MPP values are returned along with the name of the run directory. If the command fails, the name of the failing return code is returned.

When remote command monitoring is enabled, the port number is displayed at the top of the GUI.