Make port/interval changes take effect on OK (#Issue #198)
Modified the rcmd_apply() method to handle the case where remote commands are already enabled, and the user changes the port number and/or the polling interval. Previously, such a change would not take effect when the user clicked OK. They would only take effect if the GUI was closed or if the user disabled and then re-enabled remote commands.

Now, in this case, the remote command monitor is canceled and restarted with the new port and/or interval.