Explicitly specify PDF backend for matplotlib
Pyinstaller 5+ (thankfully) no longer defaults to loading all matplotlib backends. But for it to auto-detect the backend that we need, it has to be explicitly called out in the Python code.

This change imports the matplotlib "use" function and calls it with "pdf".

See https://pyinstaller.org/en/v5.0/hooks-config.html#matplotlib-hooks