Include inst path in copy destination (Issue #203)
Modify the get_dest_dir() method to insert the "inst/<name>" part of the source path in the destination path, if it exists.

Take advantage of the fact that os.path.join inserts directory separators only for non-empty segments. By initializing inst_path and overlays to "", this avoids multiple branches and should have been done previously.