Improve EMR advanced current calibration (Issue #206)
See Issue #206

Renamed the request_ssr_adv_current_calibration_val method to req_adv_current_calibration_val, but the only change is to suppress checking the cooling off period if the relay type is not SSR/FET (had to add a parameter).

When the user selects EMR on the advanced current calibration dialog, the req_adv_current_calibration_val method is now called regardless of relay type. The help dialog is consolidated to a single AdvCurrentCalHelpDialog class. It has an added step 0, which is for the EMR jumper wire, but the other steps are unchanged.