Restore full-resolution User Guide
I used Word's "Reduce File Size" option before committing the updated User Guide. But some of the figures are a bit blurry now. Although the Word doc was reduced by over a factor of two, the PDF was only reduced from 6 MB to 5.1 MB.

I decided to restore the full resolution Word doc and PDF. I will not update release v2.8.0, however. This does mean that the PDF that comes up from the Help menu will be the low res version, but that's ok. Next release will be back to the higher resolution.

There are no content changes in this commit. The only differences are the image resolutions.