Improve fix for NoSectionError exception (Issue #211)
The previously committed fix for Issue #211 works but isn't really right. The parent attribute values are not related to a run displayed in the Wizard. The reason that a simulator run doesn't have the Remote Command section in its config file is that that section is added in the GUI's extension of the base Configuration class, but the simulator uses the base class. This is the same as the Looping section. The populate_looping_tab method handles this by checking for the existence of the Looping section before applying the config values to the widgets. That is the correct way to handle this, and this commit now does the same thing in the populate_rcmd_tab method.