Create app_data_dir in check_app_data_dir method (Issue #212)
When the instances feature was added, the sequence in the GraphicalUserInterface class __init__ method was changed. The check_app_data_dir() method is now called before the ivs2 object is created. The ivs2 object creation creates the app_data_dir folder if it doesn't exist. Since it is now called after the call to check_app_data_dir(), the creation of a dummy file in that folder fails with the FATAL ERROR for anyone that had not run the app before.

It really didn't make sense to check that it was possible to create a folder in the app_data_dir parent after the IV_Swinger2 folder had already been created anyway.

The fix is to create the app_data_dir folder in the check_app_data_dir() method if it doesn't already exist. It does this in place of creating (and removing) a dummy folder.