Convert app_data_dir to Path (Issue #210)
This is one of many incremental changes to refactor the code to use pathlib Path objects. In this commit, the app_data_dir path is converted, as well as paths that have app_data_dir immediately on the RHS of their assignment. It only goes that far, though.

This commit is to the pathlib_refactor branch and will be merged to the default branch after all the other changes have been committed and thorough testing has been performed.