Convert hdd_output_dir to Path (Issue #210)
This is another incremental change to refactor the code to use pathlib Path objects. In this commit, the hdd_output_dir path is converted, as well as paths that have it immediately on the RHS of their assignment. There are a few instances where this is propagated a bit further, but for the most part that is as far as it goes..

The commit also includes a change to remove the gen_dbg_str global function from IV_Swinger2.py. It is superseded by the print_dbg_str global function in IV_Swinger.py. That function is aliased to PRINT_DBG_STR in IV_Swinger2.py, IV_Swinger2_gui.py, and IV_Swinger2_sim.py (so far).

This commit is to the pathlib_refactor branch and will be merged to the default branch after all the other changes have been committed and thorough testing has been performed.