Next incremental batch of pathlib changes (Issue #210)
This one converts the last few os.path usages in IV_Swinger_plotter.py to pathlib equivalents. It also adds the --plot_dir option to the plotter, which allows the removal of the os.chdir() call in the run() method of the IV_Swinger2_plotter class in IV_Swinger2.py.