Fix import to instance (Issue #217)
import_results() method: If import is to an instance, set the copy destination to the import folder itself instead of its parent.
get_dest_dir() method: if destination is an instance, don't add the app name or source directory instance path.
display_copy_summary() method: display the correct path (don't append IV_Swinger2) if the destination is an instance